% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RPackagesWeek4-package.R
\docType{package}
\name{RPackagesWeek4-package}
\alias{RPackagesWeek4}
\alias{RPackagesWeek4-package}
\title{RPackagesWeek4: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
